Revert "Merge tag 'v4.19-rc6'"
This reverts commit 3a5171c5ce6f24d4f00d7fc8b0c2e1e898262a4c, reversing
changes made to dcdc0bbf250cb04b750226bc7fbb3fd473b41ace.

+++ drivers/acpi/acpica/rsdump.c
@@ -539,7 +539,7 @@ static void acpi_rs_out_title(const char *title)
 
 static void acpi_rs_dump_byte_list(u16 length, u8 * data)
 {
-	u16 i;
+	u8 i;
 
 	for (i = 0; i < length; i++) {
 		acpi_os_printf("%25s%2.2X : %2.2X\n", "Byte", i, data[i]);
